app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Smorest
flask-restx
flask_migrate
Marshmallow
Werkzeug
PyJWT
requests
